+++ src/test/ts/browser/EditorBehaviorTest.ts
@@ -5,7 +5,8 @@ import { describe, it } from '@ephox/bedrock-client';
 import { getHugeRTE } from '../../../main/ts/HugeRTE';
 import { EventStore, VERSIONS } from '../alien/TestHelpers';
 import { Editor as HugeRTEEditor, EditorEvent, Events } from 'hugerte';
-import { Assertions } from '@ephox/agar';
+import { Assertions, Waiter } from '@ephox/agar';
+import { TinySelections } from '@ephox/mcagar'; // TODO get rid of it
 
 type SetContentEvent = EditorEvent<Events.EditorEventMap['SetContent']>;
 
@@ -83,6 +84,7 @@ describe('EditorBehaviourTest', () => {
         await ctx.reRender({ onSetContent: eventStore.createHandler('onSetContent') });
 
         Assertions.assertHtml('Checking HugeRTE content', '<p>Initial Content</p>', ctx.editor.getContent());
+        await Waiter.pWait(0); // Wait for React's state updates to complete before setting new content
         ctx.editor.setContent('<p>New Content</p>');
 
         eventStore.each<SetContentEvent>('onSetContent', (events) => {
@@ -110,6 +112,7 @@ describe('EditorBehaviourTest', () => {
         ctx.editor.setContent('<p>Initial Content</p>');
 
         await ctx.reRender({ onSetContent: eventStore.createHandler('NewHandler') });
+        await Waiter.pWait(0); // Wait for React's state updates to complete before setting new content
         ctx.editor.setContent('<p>New Content</p>');
 
         eventStore.each<SetContentEvent>('InitialHandler', (events) => {
@@ -128,6 +131,23 @@ describe('EditorBehaviourTest', () => {
         });
 
         eventStore.clearState();
+      });
+      it('INT-3226: onEditorChange is triggered only once after calling insertContent', async () => {
+        using ctx = await render({ onEditorChange: eventStore.createHandler('onEditorChange') });
+        const { editor } = ctx;
+        editor.setContent('<p>abc</p>');
+        await Waiter.pTryUntilPredicate('Editor content is set to correct value', () => ctx.editor.getContent() === '<p>abc</p>');
+        eventStore.clearState();
+        TinySelections.setSelection(editor, [ 0, 0 ], 1, [ 0, 0 ], 2);
+        editor.insertContent('e');
+        await Waiter.pTryUntilPredicate('Editor content is set to correct value', () => ctx.editor.getContent() === '<p>aec</p>');
+        eventStore.each<string>('onEditorChange', (events) => {
+          Assertions.assertEq(
+            'onEditorChange should have been triggered once',
+            1,
+            events.length
+          );
+        });
       });
     })
   );