+++ src/test/ts/browser/EditorInitTest.ts
@@ -45,7 +45,7 @@ describe('EditorInitTest', () => {
 
         it('gets set automatically to uuid if not set', async () => {
           using ctx = await render();
-          Assertions.assertEq('Should not be uuid', typeof ctx.DOMNode.id === 'string' && ctx.DOMNode.id.indexOf('tiny-react') !== -1, true);
+          Assertions.assertEq('Should not be uuid', typeof ctx.DOMNode.id === 'string' && ctx.DOMNode.id.indexOf('hugerte-react') !== -1, true);
         });
       });
 
@@ -64,14 +64,14 @@ describe('EditorInitTest', () => {
       it('Value prop should propagate changes to editor', async () => {
         using ctx = await render({ value: '<p>Initial Value</p>' });
         Assertions.assertHtml('Checking HugeRTE content', '<p>Initial Value</p>', ctx.editor.getContent());
-        ctx.reRender({ ...defaultProps, value: '<p>New Value</p>' });
+        await ctx.reRender({ ...defaultProps, value: '<p>New Value</p>' });
         Assertions.assertHtml('Checking HugeRTE content', '<p>New Value</p>', ctx.editor.getContent());
       });
 
       it('Disabled prop should disable editor', async () => {
         using ctx = await render();
         Assertions.assertEq('Should be design mode', true, ctx.editor.mode.get() === 'design');
-        ctx.reRender({ ...defaultProps, disabled: true });
+        await ctx.reRender({ ...defaultProps, disabled: true });
         Assertions.assertEq('Should be readonly mode', true, ctx.editor.mode.get() === 'readonly');
       });