+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The HugeRTE React component 1.0.0 is based on the TinyMCE React component 6.0.0-rc. Unreleased changes from Tiny since 5.1.0 are included in the 1.0.0 changelog.
 
+## 2.0.0 - 2025-06-30
+
+### Removed
+- Removed `propTypes` property on the Editor class.
+- Removed `prop-types` dependency and some dev dependencies.
+- Removed the storybook stuff – a new demo will be added soon.
+- Removed an IE/old Edge specific fallback code.
+
+### Changed
+- Renamed the prefix of the default editor wrapper element ID from `tiny-react` to `hugerte-react`.
+- Bumped some dependencies, including `hugerte`.
+
+### Added
+- Added support for React 19.x.
+- Added support for `onResizeEditor` event.
+
+## Fixed
+- The onEditorChange callback was called three times when content was inserted with insertContent editor API. (Not sure if this bug did exist in 1.x, but anyway tests verify it's not present now.)
+
 ## 1.0.1 - 2025-01-11
 
 ### Fixed